feat: continuous work on user sub menu and move user profile picture to submenu (WIP), refs #508
