+++ client/src/app/components/userpreferences/user-submenu/UserSubmenu.scss
@@ -10,7 +10,7 @@
   .menu-container {
     display: flex;
     flex-direction: column;
-    height: 200px;
+    // height: 200px;
     width: 300px;
     border: 1px solid red;
     padding: 20px;
