+++ client/src/scss/_user-preference.scss
@@ -6,6 +6,35 @@
   height: 100vh;
   width: 100vw;
 
+  // .right-menu-slide-right {
+  //   -webkit-animation: slide-right 0.5s cubic-bezier(0.250, 0.460, 0.450, 0.940) both;
+  //   animation: slide-right 0.5s cubic-bezier(0.250, 0.460, 0.450, 0.940) both;
+  // }
+
+  @-webkit-keyframes right-menu-slide-right {
+    0% {
+      -webkit-transform: translateX(0);
+      transform: translateX(0);
+    }
+
+    100% {
+      -webkit-transform: translateX(100px);
+      transform: translateX(100px);
+    }
+  }
+
+  @keyframes right-menu-slide-right {
+    0% {
+      -webkit-transform: translateX(0);
+      transform: translateX(0);
+    }
+
+    100% {
+      -webkit-transform: translateX(100px);
+      transform: translateX(100px);
+    }
+  }
+
   .user-setting-picture {
     display: flex;
     flex-direction: column;
